+++ src/components/fundraiser/FundraiserProgress.tsx
@@ -4,8 +4,7 @@ export const FundraiserProgress: FC<{
   raisedTitle: string
   currency: string
   raised: number
-  goal: number
-}> = ({ currency, goal, raised, raisedTitle }) => {
+}> = ({ currency, raised, raisedTitle }) => {
   const moneyFormatter = new Intl.NumberFormat(undefined, {
     style: 'currency',
     currency: currency,
@@ -13,27 +12,10 @@ export const FundraiserProgress: FC<{
     minimumFractionDigits: 0,
   })
 
-  const progressPercent = (raised / goal) * 100
-
   return (
-    <div className="prose text-navy-200">
-      <div className="mb-4 py-1 px-1 bg-navy-400">
-        <div
-          className="bg-navy-900"
-          style={{
-            width: `${progressPercent}%`,
-            height: '24px',
-          }}
-        ></div>
-      </div>
-      <div className="flex justify-between uppercase leading-tight">
-        <div>
-          {raisedTitle}
-          <br />
-          Funded: {moneyFormatter.format(raised)}
-        </div>
-        <div>Target: {moneyFormatter.format(goal)}</div>
-      </div>
-    </div>
+    <p className="progress">
+      <span className="raised">{raisedTitle}:</span>{' '}
+      <span className="raised money">{moneyFormatter.format(raised)}</span>
+    </p>
   )
 }
